Fix DoPaint method of GraphDocument was called repeatedly because GraphDocument was not suspended during paint
Fix problem with stored event args:
This problem occurs when a parent node is suspended, and a child node is also suspended manually. When the child node resumes, the accumulated event args are tried to send to the parent, but because the
parent is still suspended, it will refuse the event args, and thus this event args were lost. I added code that sets back the event args if such a case happens.

Fix that Altaxo is not brought to the foreground if it is activated by clicking on an embedded document.