Fix sporadic exception in DataColumnCollection.RemoveColumnsAll
RemoveColumnsAll calls RemoveColumns to remove a range of columns at one. During removal, the Dispose function of the removed column was called, which triggers some events (in all listerners, here: DataTableMatrixProxy). The listener then maybe try
to access the DataColumnCollection, but this is now still in an undefined state.
Fixed this by collection of the columns to dispose and dispose them after the DataColumnCollection is again in a defined state.