Allow for a number of trials when the rendering fails with an exception
- Altaxo is not thread safe. Thus it can happen that during the background rendering of a graph the graph itself or
the underlying data changed, sometimes causing a rendering exception. If that happens, we try to render again, hoping that now the data have settled or the graph isn't changing any more. If an exception is thrown again, we repeat this until either a given number of trials is exceeded, or a given time has expired.