Fix bug when copying a project item with relocation that had references to folders below
- because DocNodeReplacementOptions stored only replacements for the immediate source and destination, a reference to a folder below was not relocated. Example:
The worksheet is Foo/WKS0, the Graph that plots data from this worksheet in Foo/Bar/Graph0
There is a second worksheet in folder Box: Box/WKS0
Trying to copy the graph Foo/Bar/Graph0 to Box/Bar/Graph0 with relocate options doesn't relocate the reference from
Foo/WKS0 to Box/WKS0, as expected. This was caused because replacement options were stored only for Foo/Bar to Box/Bar, but not from Foo to Box. Resolved by changing the replacement dictionary to a multilevel dictionary, with more priority for the immediate folder replacements, then for the stripped-off folders.